I'm not sure what I changed here
I want to use the built in PID position control but that means we lose
the efficiency of moving from 360 to 10. Options:
1. Somehow figure out the optimal position at all times that goes beyond
360. Then easily implement the PID position control. Actually this might
not be too hard. It's just find the target between 0-360, compare it to
the current value, multiply the target until it's as close as possible
to the target. I have to think about this
2. Rewrite the Jason method so it behaves like a relatively good PID.
Dr. Bayntun thanks 
3. Give up on Swerve and give up on life